putting up a functional server environment
