+++ package.json
@@ -2,7 +2,7 @@
   "name": "tictactoe-hug",
   "version": "0.0.1",
   "description": "Late Term Assignment for hugbunadarfraedi",
-  "main": "app.js",
+  "main": "server.js",
   "dependencies": {
     "express": "^4.16.4",
     "jquery": "^3.3.1"
@@ -22,7 +22,7 @@
     "test": "jest",
     "watch": "jest --watchAll",
     "clean": "rm -rf node_modules",
-    "start": "node app.js",
+    "start": "node server.js",
     "build": "webpack",
     "wds": "webpack-dev-server"
   },
