--- src/server/api.js
@@ -1,7 +1,0 @@
-// src/server/api.js
-const express = require("express");
-const router = express.Router();
-router.get("/", (req, res) => {
-   res.status(200).send('Hello World');
-});
-   module.exports = router;
